requirements.txt
PyYAML>=6.0.2
deepmerge>=2.0
argparse